refactor : 원래대로 수정
